requirements.txt
# Drug-Drug Interaction Prediction System v2.0 - Requirements

# Core ML and Data Processing
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Chemistry and Molecular Processing
rdkit==2023.9.3

# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
google-generativeai==0.3.2
python-dotenv==1.0.0
email-validator>=2.1.0
